% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Score.R
\name{Risk_Score}
\alias{Risk_Score}
\alias{Risk_Score.vector}
\alias{Risk_Score.data.frame}
\title{Risk Score}
\usage{
Risk_Score(diabetes, gender, smoke, age, sbp, cholesterol)

\method{Risk_Score}{vector}(diabetes, gender, smoke, age, sbp, cholesterol)

\method{Risk_Score}{data.frame}(df)
}
\arguments{
\item{diabetes}{Binary form, non-diabetic = 0; diabetic = 1.}

\item{gender}{Binary form, 0 = female; 1 = male.}

\item{smoke}{Binary form, 0 = non-smoker; 1 = smoker.}

\item{age}{Age in years}

\item{sbp}{Systolic Blood Pressure in mmHG}

\item{cholesterol}{Total cholesterol in mmol/l}

\item{df}{A dataframe with columns in the following order: Diabetes (0 = non-diabetic, 1 = diabetic), Gender (0 = female, 1 = male), Smoke (0 = non-smoker, 1 = smoker), age (year) , sbp (Sytolic Blood Pressure in mmHG), cholesterol (Total cholesterol in mmol / l)}
}
\value{
A data.frame with the Score of risk level of Cardiovascular Disease

\code{NULL}

\code{NULL}
}
\description{
A function that return the risk level of Cardiovascular Disease in Central Latin America. Based on WHO cardiovascular disease risk.
}
\examples{
Risk_Score(diabetes = 1, gender = 0, smoke = 1, age = 60,

Age <- c(44, 57, 66, 50, 71, 37)
Gender <- c(0,1,0,1,1, 1)
Smoking <- c(1,1,0,1,0, 1)
Systolic_Blood_Pressure <- c(120, 137, 144, 178, 163, 139)
Diabetes <- c(1,1,1,0,1, 1)
cholesterol <- c(0, 5.8, 6.5, 0, 7.2, 6)
Risk_Score.vector(diabetes = Diabetes, gender = Gender,smoke = Smoking, age = Age, sbp = Systolic_Blood_Pressure,cholesterol = cholesterol)

Age <- c(44, 57, 66, 50, 71, 37)
Gender <- c(0,1,0,1,1, 1)
Smoking <- c(1,1,0,1,0, 1)
Systolic_Blood_Pressure <- c(120, 137, 144, 178, 163, 139)
Diabetes <- c(1,1,1,0,1, 1)
cholesterol <- c(0, 5.8, 6.5, 0, 7.2, 6)
df <- data.frame(Diabetes, Gender, Smoking, Age, Systolic_Blood_Pressure, cholesterol)

Risk_Score(df)

}
\author{
Erick Cuevas-Fernández
}
