% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rick_ScoreBMI.R
\name{Risk_ScoreBMI}
\alias{Risk_ScoreBMI}
\alias{Risk_ScoreBMI.vector}
\alias{Risk_ScoreBMI.data.frame}
\title{Risk Score Body Mass Index}
\usage{
Risk_ScoreBMI(gender, smoke, age, sbp, bmi)

\method{Risk_ScoreBMI}{vector}(gender, smoke, age, sbp, bmi)

\method{Risk_ScoreBMI}{data.frame}(df)
}
\arguments{
\item{gender}{Binary form, 0 = female; 1 = male.}

\item{smoke}{Binary form, 0 = non-smoker; 1 = smoker.}

\item{age}{Age in years}

\item{sbp}{Systolic Blood Pressure in mmHG}

\item{bmi}{Body Mass Index (kg/m^2)}

\item{df}{A dataframe with columns in the following order: Gender (0 = female, 1 = male), Smoke (0 = non-smoker, 1 = smoker), age (year) , sbp (Sytolic Blood Pressure in mmHG), BMI (Body Mass Index in kg/m^2)}
}
\value{
A data.frame with the Score of risk level of Cardiovascular Disease

\code{NULL}

\code{NULL}
}
\description{
A function that return the risk level of Cardiovascular Disease in Central Latin America. Based on WHO cardiovascular disease risk.
}
\examples{
Risk_ScoreBMI(gender = 0, smoke = 1, age = 60,

Age <- c(44, 57, 66, 50, 71, 37)
Gender <- c(0,1,0,1,1, 1)
Smoking <- c(1,1,0,1,0, 1)
Systolic_Blood_Pressure <- c(120, 137, 144, 178, 163, 139)
BMI <- c(20,34,30,27,21, 19)
Risk_ScoreBMI(gender = Gender,smoke = Smoking, age = Age, sbp = Systolic_Blood_Pressure,bmi = BMI)

Age <- c(44, 57, 66, 50, 71, 37)
Gender <- c(0,1,0,1,1, 1)
Smoking <- c(1,1,0,1,0, 1)
Systolic_Blood_Pressure <- c(120, 137, 144, 178, 163, 139)
bmi <- c(20,34,30,27,21, 19)
df <- data.frame(Gender, Smoking, Age, Systolic_Blood_Pressure, bmi)

Risk_ScoreBMI(df)

}
\author{
Erick Cuevas-Fernández
}
